requirements.txt
tqsdk==3.9.1
pandas==3.0.1
matplotlib==3.10.8
mplfinance==0.12.10b0
python-dotenv==1.2.2
Flask==3.1.0
